Fix http and gateway feature
Previously, you couldn't even build serenity if you had just http or just gateway enabled. And, gateway used to depend on http and utils which is non sensical. Both these things are fixed now
