+++ src/builder/edit_profile.rs
@@ -70,8 +70,7 @@ impl EditProfile {
         http: impl AsRef<Http>,
         avatar: impl Into<AttachmentType<'a>>,
     ) -> Result<Self> {
-        let avatar_data = avatar.into().data(&http.as_ref().client).await?;
-        self.avatar = Some(Some(crate::utils::encode_image(&avatar_data)));
+        self.avatar = Some(Some(avatar.into().to_base64(&http.as_ref().client).await?));
         Ok(self)
     }
