+++ src/model/channel/mod.rs
@@ -51,6 +51,7 @@ pub enum Channel {
     Private(PrivateChannel),
 }
 
+#[cfg(feature = "model")]
 impl Channel {
     /// Converts from [`Channel`] to `Option<GuildChannel>`.
     ///
@@ -124,7 +125,6 @@ impl Channel {
     ///
     /// Otherwise will return [`Error::Http`] if the current user does not
     /// have permission.
-    #[cfg(feature = "http")]
     pub async fn delete(&self, cache_http: impl CacheHttp) -> Result<()> {
         match self {
             Self::Guild(public_channel) => {
